+++ src/common/utils/index.ts
@@ -48,8 +48,9 @@ export function git(options: any = { output: false }): SimpleGit {
     function logCommand() {
       if (first) {
         first = false;
-        if (!(gitArgs && gitArgs[0] && gitArgs[0] === "branch" && gitArgs[1] && gitArgs[1] === "-v")) {
-          uxLog(this, `[command] ${c.grey(command)} ${c.grey(gitArgs.join(" "))}`);
+        const gitArgsStr = (gitArgs || []).join(" ");
+        if (!(gitArgsStr.includes("branch -v") || gitArgsStr.includes("config --list --show-origin --null"))) {
+          uxLog(this, `[command] ${c.bold(c.bgWhiteBright(c.grey(command + " " + gitArgsStr)))}`);
         }
       }
     }
@@ -502,7 +503,7 @@ export async function execCommand(
     spinner: true,
   }
 ): Promise<any> {
-  const commandLog = `[sfdx-hardis][command] ${c.bold(c.grey(command))}`;
+  const commandLog = `[sfdx-hardis][command] ${c.bold(c.bgWhiteBright(c.grey(command)))}`;
   let commandResult = null;
   // Call command (disable color before for json parsing)
   const prevForceColor = process.env.FORCE_COLOR;